fix: use blocking io to fix privileged mount hang
